refactor: Querydsl 기반의 Auction 인터페이스 제거 및 구현체에서 일반 클래스로 변환
