+++ backend/ddang/src/test/java/com/ddang/ddang/auction/infrastructure/persistence/fixture/AuctionForListSortByAuctioneerCountFixture.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Order;
 
 @SuppressWarnings("NonAsciiCharacters")
-public class AuctionForListSortByAuctioneerCountFixture extends QuerydslAuctionRepositoryImplForListFixture {
+public class AuctionForListSortByAuctioneerCountFixture extends QuerydslAuctionRepositoryForListFixture {
 
     protected Sort 참여_인원순_정렬 = Sort.by(Order.asc("auctioneerCount"));
     protected ReadAuctionSearchCondition 검색어_없음 = new ReadAuctionSearchCondition(null);
