+++ backend/ddang/src/test/java/com/ddang/ddang/auction/infrastructure/persistence/fixture/AuctionForListSortByClosingTimeFixture.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Order;
 
 @SuppressWarnings("NonAsciiCharacters")
-public class AuctionForListSortByClosingTimeFixture extends QuerydslAuctionRepositoryImplForListFixture {
+public class AuctionForListSortByClosingTimeFixture extends QuerydslAuctionRepositoryForListFixture {
 
     protected Sort 마감_임박순_정렬 = Sort.by(Order.asc("closingTime"));
     protected ReadAuctionSearchCondition 검색어_없음 = new ReadAuctionSearchCondition(null);
